fixed vphysics compile
took me only 7 months

+++ common/lzma/lzma.h
@@ -30,7 +30,7 @@
 unsigned char *LZMA_Compress(
 unsigned char	*pInput,
 unsigned int	inputSize,
-unsigned int	*pOutputSize );
+size_t      	*pOutputSize );
 
 //-----------------------------------------------------------------------------
 // Decoding glue. Returns TRUE if succesful.
@@ -38,7 +38,7 @@ unsigned int	*pOutputSize );
 bool LZMA_Uncompress(
 unsigned char	*pInput,
 unsigned char	**ppOutput,
-unsigned int	*pOutputSize );
+size_t          *pOutputSize );
 
 //-----------------------------------------------------------------------------
 // Decoding helper, returns TRUE if buffer is LZMA compressed.
